添加.ld文件 添加.s文件，复制文件到arm...目录 添加Cmakelist文件 添加arm-gnu-toolchain-14.3.Rel1-x86_64-arm-none-eabi.tar.xz编译器文件，需要将该文件解压到opt/目录下 需要设置正确的PATH echo 'export PATH="/opt/arm-gnu-toolchain-14.3.rel1-x86_64-arm-none-eabi/bin:$PATH"' >> ~/.bashrc source ~/.bashrc 修改.h文件让其尽量少包含其他.h文件，在.c文件中包含其他.h文件 该工程已经成功编译，待验证烧录bin文件到stm32中是否能正常工作

+++ STM32F407/bsp/SPI/SPI_Control.c
@@ -1,4 +1,8 @@
 #include "SPI_Control.h"
+#include "board.h"
+#include <stddef.h>
+#include <stdio.h>
+#include <sys/cdefs.h>
 
 // 定义一个静态常量数组spiErrors，用于存储所有可能的SPI错误及其对应的错误信息
 static const Error_Code spiErrors[] =
@@ -15,7 +19,7 @@ static const Error_Code spiErrors[] =
 static uint8_t SPI_TIMEOUT_UserCallback(uint8_t errorCode)
 {
   // 查找并打印错误信息（如果找到）
-  for (int i = 0; i < SPI_ERROR_COUNT; i++)
+  for (size_t i = 0; i < SPI_ERROR_COUNT; i++)
   {
     if (spiErrors[i].errorCode == errorCode)
     {
